0.4
Initially finished bike and faulties panels.

+++ Bike/assets/php/oddOrder.php
@@ -2,13 +2,12 @@
 include 'connect.php';
         
 $faultID=$_POST['id'];
-        
         if($id=$mysqli->query("SELECT id_serwis FROM przypisania_awarii WHERE id_awarii=$faultID")){
-            $row = $id->fetch_row();
-            echo $row[0];
+            if(!is_null($id)){  
+                $row = $id->fetch_row();
+                echo $row[0];
+                 $mysqli->query("UPDATE awarie SET status=1 WHERE awarie.id_awarii=$faultID");
+                 $mysqli->query("DELETE FROM przypisania_awarii WHERE przypisania_awarii.id_awarii=$faultID");
+            }
         }
-
-             $mysqli->query("UPDATE awarie SET status=1 WHERE awarie.id_awarii=$faultID");
-             $mysqli->query("DELETE FROM przypisania_awarii WHERE przypisania_awarii.id_awarii=$faultID");
-           
 ?>
